requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
sqlmodel==0.0.22
python-multipart==0.0.9
pillow==10.4.0
python-dotenv==1.0.1
apscheduler==3.10.4
pytest==7.4.4
psycopg2-binary
httpx==0.27.2
mega.py==1.0.8
